pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "mobius-ai"
dynamic = ["version"]
description = "ONNX model definitions for GenAI using onnxscript.nn"
authors = [
    { name = "ONNX Contributors", email = "onnx-technical-discuss@lists.lfaidata.foundation" },
]
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "onnxscript>=0.6.0.dev",
    "onnx_ir>=0.1.0",
    "numpy>=1.24.0",
    "torch>=2.1.0",
    "huggingface_hub",
    "tqdm",
]

[project.scripts]
mobius = "mobius.__main__:main"

[project.optional-dependencies]
transformers = ["transformers", "safetensors"]
gguf = ["gguf>=0.10.0"]
testing = [
    "onnxruntime-easy",
    "torch",
    "transformers>=5.0",
    "safetensors",
    "onnxruntime-genai",
    "pytest-xdist",
    "pytest-timeout",
    "gguf>=0.10.0",
    "Pillow",
    "librosa",
]
ort-genai = ["onnxruntime-genai"]
docs = ["furo", "myst-parser", "sphinx", "sphinx_copybutton"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "mobius.__version__" }

[tool.pytest.ini_options]
addopts = "--tb=short --color=yes"
testpaths = ["src", "tests"]
python_files = ["*_test.py"]
markers = [
    "integration: integration tests requiring model downloads and GPU/CPU inference (deselect with '-m \"not integration\"')",
    "integration_fast: integration tests for single-modality models (text, encoder, audio)",
    "integration_slow: integration tests for multi-modality models (vision-language, multimodal)",
    "benchmark: graph construction performance benchmarks",
    "arch_validation: L2 architecture validation tests that download real HF configs (no weights) and build full-size ONNX graphs (deselect with '-m \"not arch_validation\"')",
    "golden: L4 checkpoint-verified golden comparison tests (deselect with '-m \"not golden\"')",
    "generation: L5 generation end-to-end golden tests (deselect with '-m \"not generation\"')",
]

[tool.mypy]
check_untyped_defs = true
disable_error_code = 'override,import-untyped'
disallow_any_generics = false
disallow_untyped_decorators = true
show_column_numbers = true
strict_optional = true
warn_incomplete_stub = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = false

[tool.ruff]
line-length = 95
target-version = "py310"

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "D",    # pydocstyle
    "E",    # pycodestyle
    "F",    # Pyflakes
    "G",    # flake8-logging-format
    "I",    # isort
    "ISC",  # flake8-implicit-str-concat
    "LOG",  # flake8-logging
    "N",    # pep8-naming
    "NPY",  # modern numpy
    "PERF", # Perflint
    "PIE",  # flake8-pie
    "PYI",  # flake8-pyi
    "RUF",  # Ruff-specific rules
    "SIM",  # flake8-simplify
    "SLOT", # flake8-slot
    "T10",  # flake8-debugger
    "TID",  # Disallow relative imports
    "TRY",  # flake8-try-except-raise
    "UP",   # pyupgrade
    "W",    # pycodestyle
    "YTT",  # flake8-2020
]
preview = true
ignore = [
    "B9",
    "C408",
    "D1",
    "D401",
    "D417",
    "E1",
    "E2",
    "E3",
    "E501",
    "SIM103",
    "NPY002",
    "PERF203",
    "PERF401",
    "PYI041",
    "RUF022",
    "RUF031",
    "RUF052",
    "RUF067",
    "SIM102",
    "SIM108",
    "SIM114",
    "SIM116",
    "TRY003",
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pydocstyle]
convention = "google"
